Securing REST - exception handling
