+++ Beauty Manager Service/src/main/java/com/simo333/beauty_manager_service/security/jwt/AuthEntryPointJwt.java
@@ -1,6 +1,7 @@
 package com.simo333.beauty_manager_service.security.jwt;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
